add ability to select another image

+++ server/db/index.js
@@ -9,6 +9,9 @@ Page.belongsTo(Image);
 Image.hasMany(Page);
 
 const syncAndSeed = async()=> {
+  if(process.env.SYNC !== 'true'){
+    return;
+  }
   await db.sync({ force: true });
   const User = models.User;
   let _users = [
@@ -22,6 +25,7 @@ const syncAndSeed = async()=> {
     })
   ];
   if(process.env.ADMIN_EMAIL && process.env.ADMIN_PASSWORD){
+    //if user exists do not recreate?
     _users = [
       User.create({
         email: process.env.ADMIN_EMAIL,
@@ -30,6 +34,7 @@ const syncAndSeed = async()=> {
     ]
   }
   const users = await Promise.all(_users);
+  console.log('seeded');
 
   return users.reduce((acc, user)=> {
     acc[user.email] = user;
